refactor: consolidate toErrorMessage/ensureError into @common/errors
`toErrorMessage` and `ensureError` were implemented in
`src/utils/core/errorCore.ts` but surfaced through two different
public barrels — `@utils/core` and `@common/errors` — with no
documented rule for which to use. All ~110 callers already import
from `@common/errors`; the `@utils/core` re-export was dead weight
that could mislead future contributors into using the wrong path.

- Move the implementation directly into `src/common/errors/errorMessage.ts`
  (previously a one-line re-export shim)
- Delete `src/utils/core/errorCore.ts`
- Remove the stale `toErrorMessage`/`ensureError` exports from the
  `@utils/core` barrel (`src/utils/core/index.ts`)
- Fix the only caller that used `@utils/core` for `toErrorMessage`
  (`packages/extension/src/progressView/frontend/formatters/baseLogFormatter.ts`)
- Update stale docstrings in `stringCore.ts` that pointed to `@utils/core`

No functional change — the implementation is identical; only the
ownership and import path are clarified.
